requirements.txt
google-generativeai
pymupdf
arxiv
faiss-cpu
sentence-transformers
